Composer: Force some files to load.
We need `Autoloader.php` to be always present in order to:
  1. Resolve flex-entities,
  2. Have a default autoloader in case Composer is not present.

Well, the second point is a little bit absurb because if Composer is
not present, this file might be useless, except if some tools (including
us) are based on this file.

We also need `Consistency.php` because this is the consistency layer. It
must be always available, to ensure foreward and backward compatibility.