Prelude: Introduce the prelude/preamble!
The last missing piece to remove `Hoa\Core` totally is… a prelude! This
makes sense that `Hoa\Consistency` brings the prelude since this library
is required by all the others (for compatibility purposes but also
because of the autoloader).

Actually, `Hoa\Core` contains clearly separated libraries, namely
`Hoa\Event`, `Hoa\Consistency`, `Hoa\Exception` etc. but a last piece of
code remained and it appears this is strictly a prelude, like we have in
Rust or C: A set of functionnalities that is always provided and
presents. Consequently, instead of loading `Hoa/Core/Core.php`, we just
need to load `Hoa/Consistency/Prelude.php`, which makes a lot of sense.