Autoloader: Restrict loads to mapped entities.
This autoloader respects the PSR-4 recommendation. However, it adds
support for flex entities, i.e. if `A\B\C` does not exist, then we try
to load `A\B\C\C`. By default, we try this for all kind of entities.
This is wrong because some autoloaders in the wild can not like it.

Consequently, we restrict the flex entity algorithm to mapped entities
only. A new `hasBaseDirectory` method has been created to check if a
namespace prefix has a base directory registered or not. This method has
been extracted from `getBaseDirectories`, nothing fancy here. But this
method is used to check whether at least one base directory exists for
an entity. If yes, this is a mapped entity and we can apply the flex
entity algorithm, else, nothing to do: Fail silently and move to the
next autoloader in the stack.